Created dialog box that pops up when creating character. This box allows the user to define certain starting parameters for characters.
View: Added some of the UI images needed for this dialog box. Created the window for the dialog box. Transplanted the previous functionality from the 'New Character' button in MainWindow to a new button in the dialog.

ViewModel: created the ViewModel responsible for controlling the dialog box. Dialog is created from the MainViewModel